+++ esper/src/edu/brown/benchmark/voteresper/dataconnectors/EsperDataConnector.java
@@ -8,11 +8,11 @@ public abstract class EsperDataConnector {
 	public abstract boolean hasVoted(long phoneNumber);
 	public abstract boolean realContestant(int contestant);
 	
-	public abstract int numTimesVoted(long phoneNumber);
+	public abstract long numTimesVoted(long phoneNumber);
 	public abstract String getState(long phoneNumber);
-	public abstract int getNumVotes();
+	public abstract long getNumVotes();
 	public abstract int getAllVotesEver();
-	public abstract int getNumRemainingContestants();
+	public abstract long getNumRemainingContestants();
 	
 	public abstract boolean insertVote(Vote v);
 	public abstract boolean updateLeaderboards(Vote v);
